+++ palace/utils/configfile.cpp
@@ -56,6 +56,7 @@ PALACE_JSON_SERIALIZE_ENUM(SparseCompression)
 PALACE_JSON_SERIALIZE_ENUM(Orthogonalization)
 PALACE_JSON_SERIALIZE_ENUM(DomainOrthogonalizationWeight)
 PALACE_JSON_SERIALIZE_ENUM(Device)
+PALACE_JSON_SERIALIZE_ENUM(InactivePortMode)
 }  // namespace palace
 
 namespace palace::config
@@ -594,6 +595,10 @@ SurfaceCurrentData::SurfaceCurrentData(const json &source)
       auto &elem = elements.emplace_back();
       ParseElementData(e, true, elem);
     }
+  }
+  if (source.find("InactiveMode") != source.end())
+  {
+    inactive_port_mode = source.at("InactiveMode").get<InactivePortMode>();
   }
 }
 
@@ -1301,6 +1306,7 @@ ElectrostaticSolverData::ElectrostaticSolverData(const json &electrostatic)
 MagnetostaticSolverData::MagnetostaticSolverData(const json &magnetostatic)
 {
   n_post = magnetostatic.value("Save", n_post);
+  inactive_port_mode = magnetostatic.value("InactivePorts", inactive_port_mode);
 }
 
 TransientSolverData::TransientSolverData(const json &transient)